extract remaining model panels
