+++ src/main/java/frontend/ModelTab.java
@@ -1,35 +1,17 @@
 package frontend;
 
-import utils.Logging.MiniMLLogger;
 import workflow.Keys;
 import workflow.WorkflowManager;
 import workflow.context.AbstractCompositeContext;
 import workflow.context.ModelContext;
-import workflow.context.ParameterContext;
 
 import javax.swing.*;
-import javax.swing.event.ChangeEvent;
-import javax.swing.event.ChangeListener;
 import java.awt.*;
 
 public class ModelTab extends JComponent {
 
     private AbstractCompositeContext parentContext;
     private ModelContext context;
-    private ParameterContext methodSelectContext;
-    private ParameterContext ridgeContext;
-    private ParameterContext hiddenLayerContext;
-    private ParameterContext hiddenNodesContext;
-    private ParameterContext learnRateContext;
-    private ParameterContext epochContext;
-    private ParameterContext gammaContext;
-    private ParameterContext epsilonContext;
-    private ParameterContext degreeContext;
-    private ParameterContext nuContext;
-    private ParameterContext pruningConfidenceContext;
-    private ParameterContext pruningFoldsContext;
-    private ParameterContext decisionTreeInstancesContext;
-
 
     public ModelTab(){
         super();
@@ -40,12 +22,13 @@ public class ModelTab extends JComponent {
         this.setLayout(new GridLayout());
         JPanel panel = new JPanel(false);
         panel.setLayout(new GridLayout(6,0));
+
         JPanel etaPanel = new ModelTabEtaPanel();
         JPanel algorithmPanel = new ModelTabAlgorithmPanel();
-        JPanel linearRegressionPanel = getLinearRegressionPanel();
-        JPanel neuralNetPanel = getNeuralNetPanel();
-        JPanel supportVectorMachinePanel = getSupportVectorMachinePanel();
-        JPanel decisionTreePanel = getDecisionTreePanel();
+        JPanel linearRegressionPanel = new ModelTabLrPanel();
+        JPanel neuralNetPanel = new ModelTabNnPanel();
+        JPanel supportVectorMachinePanel = new ModelTabSvmPanel();
+        JPanel decisionTreePanel = new ModelTabDtPanel();
 
         panel.add(etaPanel);
         panel.add(algorithmPanel);
@@ -56,303 +39,4 @@ public class ModelTab extends JComponent {
 
         this.add(panel);
     }
-
-    private JPanel getLinearRegressionPanel(){
-        this.methodSelectContext = new ParameterContext(this.context, Keys.MethodSelect);
-        this.ridgeContext = new ParameterContext(this.context, Keys.Ridge);
-
-        JPanel panel = new JPanel();
-        String val1 = "";
-        String val2 = "";
-        panel.setLayout(new GridLayout());
-        JLabel tip = new JLabel("Linear Regression: ");
-        JProgressBar pbar = new JProgressBar(0,0,100); // Do we need this?
-        JLabel par1 = new JLabel("Select Method: " + val1);
-        JLabel par2 = new JLabel("Ridge: " + val2);
-        JSlider slider1 = new JSlider(0,2,1);
-        JSlider slider2 = new JSlider(0, 1, 1);
-
-        slider1.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.MethodSelect);
-                    handleContext((ParameterContext) context, slider1.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider2.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.Ridge);
-                    handleContext((ParameterContext) context, slider2.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-
-        panel.add(tip);
-        panel.add(par1);
-        panel.add(slider1);
-        panel.add(par2);
-        panel.add(slider2);
-        return panel;
-    }
-
-    private JPanel getNeuralNetPanel(){
-        this.hiddenLayerContext = new ParameterContext(this.context, Keys.NumHiddenLayers);
-        this.hiddenNodesContext = new ParameterContext(this.context, Keys.NumHiddenNodes);
-        this.learnRateContext = new ParameterContext(this.context, Keys.LearnRate);
-        this.epochContext = new ParameterContext(this.context, Keys.NumEpochs);
-
-        JPanel panel = new JPanel();
-        String val1 = "";
-        String val2 = "";
-        String val3 = "";
-        String val4 = "";
-        panel.setLayout(new GridLayout());
-        JLabel tip = new JLabel("Neural Network: ");
-        JProgressBar pbar = new JProgressBar(0,0,100);
-        JLabel par1 = new JLabel("Hidden Layers: " + val1);
-        JLabel par2 = new JLabel("Hidden Nodes: " + val2);
-        JLabel par3 = new JLabel("Learn Rate: " + val3);
-        JLabel par4 = new JLabel("Epochs: " + val4);
-        JSlider slider1 = new JSlider(1,9,1);
-        JSlider slider2 = new JSlider(1, 100, 10);
-        JSlider slider3 = new JSlider(0,100,20);
-        JSlider slider4 = new JSlider(1, 1000, 100);
-
-        slider1.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.NumHiddenLayers);
-                    handleContext((ParameterContext) context, slider1.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider2.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.NumHiddenNodes);
-                    handleContext((ParameterContext) context, slider2.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider3.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.LearnRate);
-                    handleContext((ParameterContext) context, slider3.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider4.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.NumEpochs);
-                    handleContext((ParameterContext) context, slider4.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        panel.add(tip);
-        panel.add(par1);
-        panel.add(slider1);
-        panel.add(par2);
-        panel.add(slider2);
-        panel.add(par3);
-        panel.add(slider3);
-        panel.add(par4);
-        panel.add(slider4);
-        return panel;
-    }
-
-    public JPanel getSupportVectorMachinePanel(){
-        this.gammaContext = new ParameterContext(this.context, Keys.Gamma);
-        this.epsilonContext = new ParameterContext(this.context, Keys.Epsilon);
-        this.degreeContext = new ParameterContext(this.context, Keys.Degree);
-        this.nuContext = new ParameterContext(this.context, Keys.Nu);
-
-        JPanel panel = new JPanel();
-        String val1 = "";
-        String val2 = "";
-        String val3 = "";
-        String val4 = "";
-        panel.setLayout(new GridLayout());
-        JLabel tip = new JLabel("Support Vector Machine: ");
-        JProgressBar pbar = new JProgressBar(0,0,100);
-        JLabel par1 = new JLabel("Gamma: " + val1);
-        JLabel par2 = new JLabel("Epsilon: " + val2);
-        JLabel par3 = new JLabel("Degree: " + val3);
-        JLabel par4 = new JLabel("Nu: " + val4);
-        JSlider slider1 = new JSlider(1,9,1);
-        JSlider slider2 = new JSlider(1, 100, 10);
-        JSlider slider3 = new JSlider(0,100,20);
-        JSlider slider4 = new JSlider(1, 1000, 100);
-
-        slider1.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.Gamma);
-                    handleContext((ParameterContext) context, slider1.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider2.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.Epsilon);
-                    handleContext((ParameterContext) context, slider2.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider3.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.Degree);
-                    handleContext((ParameterContext) context, slider3.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider4.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.Nu);
-                    handleContext((ParameterContext) context, slider4.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        panel.add(tip);
-        //panel.add(pbar);
-        panel.add(par1);
-        panel.add(slider1);
-        panel.add(par2);
-        panel.add(slider2);
-        panel.add(par3);
-        panel.add(slider3);
-        panel.add(par4);
-        panel.add(slider4);
-        return panel;
-    }
-
-    public JPanel getDecisionTreePanel(){
-        this.pruningConfidenceContext = new ParameterContext(this.context, Keys.PruningConfidence);
-        this.pruningFoldsContext = new ParameterContext(this.context, Keys.PruningFolds);
-        this.decisionTreeInstancesContext = new ParameterContext(this.context, Keys.DecisionTreeInstances);
-
-        JPanel panel = new JPanel();
-        String val1 = "";
-        String val2 = "";
-        String val3 = "";
-        panel.setLayout(new GridLayout());
-        JLabel tip = new JLabel("Decision Tree: ");
-        JProgressBar pbar = new JProgressBar(0,0,100);
-        JLabel par1 = new JLabel("Pruning Confidence: ");
-        JLabel par2 = new JLabel("Pruning Folds: ");
-        JLabel par3 = new JLabel("Instances: ");
-        JSlider slider1 = new JSlider(1,9,1);
-        JSlider slider2 = new JSlider(1, 100, 10);
-        JSlider slider3 = new JSlider(0,100,20);
-
-        slider1.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.PruningConfidence);
-                    handleContext((ParameterContext) context, slider1.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider2.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.PruningFolds);
-                    handleContext((ParameterContext) context, slider2.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        slider3.addChangeListener(new ChangeListener() {
-            private AbstractCompositeContext context;
-            @Override
-            public void stateChanged(ChangeEvent e) {
-                try {
-                    context = WorkflowManager.INSTANCE.getContextByKey(Keys.DecisionTreeInstances);
-                    handleContext((ParameterContext) context, slider3.getValue());
-                } catch(Exception exception) {
-                    MiniMLLogger.INSTANCE.exception(exception);
-                }
-            }
-        });
-
-        panel.add(tip);
-        panel.add(par1);
-        panel.add(slider1);
-        panel.add(par2);
-        panel.add(slider2);
-        panel.add(par3);
-        panel.add(slider3);
-        return panel;
-    }
-
-    public void handleContext(ParameterContext context, Object value) {
-        context.setValue(value);
-        context.updateState();
-    }
-
 }
